refactor: move `isValidSystemType` to `document` helper module
