+++ module/data/fields/sibling-document-field.mjs
@@ -14,7 +14,7 @@
  */
 
 
-import { isValidSystemType } from "../../utils/foundry.mjs";
+import { isValidSystemType } from "../../helper/document.mjs";
 
 /**
  * A ForeignDocumentField which references a sibling document within the same parent document.
