+++ module/helper/document.mjs
@@ -1,3 +1,5 @@
+import { SYSTEM_TYPES } from "../constants/constants.mjs";
+
 /**
  * Adapted from ({@link https://gitlab.com/peginc/swade/-/wikis/Savage-Worlds-ID|SWADE system}).
  * Returns an array of items that match a given EDID and optionally an item type.
@@ -120,4 +122,23 @@ export async function getAllDocuments(
     ? allDocuments.map( doc => doc.uuid )
     : Promise.all( allDocuments.map( doc => fromUuid( doc.uuid ) ) );
 
+}
+
+/**
+ * Check whether the provided system type is valid for the given document type,
+ * or in general if no document type is provided.
+ * @param {string} systemType The system type to check.
+ * @param {string} [documentType] The document type to check against.
+ * @returns {boolean} True if the system type is valid, false otherwise.
+ */
+export function isValidSystemType( systemType, documentType ) {
+  if ( documentType && !Object.keys( SYSTEM_TYPES ).includes( documentType ) ) return false;
+
+  const validTypes = documentType
+    ? Object.values( SYSTEM_TYPES[documentType] )
+    : Object.values( SYSTEM_TYPES ).map(
+      types => Object.values( types )
+    ).flat();
+
+  return validTypes.includes( systemType );
 }
